Module Monoid.Free

The free monoid on a set.

Parameters

module X : Alphabet.T

Signature

type t = X.t array

type word = t

val is_commutative : bool

val mul : t -> t -> t

Multiplication, i.e. concatenation.

val one : t

Unit of the monoid.

val pow : t -> int -> t

val inj : X.t -> t

Canonical injection from the alphabet to the monoid.

val length : t -> int

Length of a word.

val sub : t -> int -> int -> t

val eq : t -> t -> bool

val compare : t -> t -> int

val peq : t -> int -> t -> int -> int -> bool

Partial equality (equality on subwords).

val to_string : t -> string

String representation of a word.

module Map (M : T) : sig ... end

Maps from the free monoid to another monoid.

val included : t -> t -> bool

val unifier : ?i:int -> t -> t -> int

The leftmost unifier where u is on the left and v on the right

val ordered_unifiers : t -> t -> int list

All unifiers with first on the left.

val ordered_unifiers_bicontext : t -> t -> ((t * t) * (t * t)) list

val unifiers_bicontext : t -> t -> ((t * t) * (t * t)) list

All unifiers, with contexts on the left and on the right.

module Order : sig ... end

module Anick : sig ... end

Anick chains.