Module RS.Zigzag

Rewriting zigzags.

type t =

| Step of Step.t

| Comp of t * t

| Id of term

| Inv of t

A rewriting zigzag.

val to_string : ?pa:bool -> ?var:( Var.t -> string ) -> t -> string

String representation.

val source : t -> term

val target : t -> term

val step : Step.t -> t

Path reduced to one step.

val comp : t -> t -> t

Concatenation of two paths.

val append : t -> t -> t

val concat : t list -> t

Concatenation of a list of paths.

val inv : t -> t

Inverse of a path.

val eq : t -> t -> bool

Equality between paths.

val length : t -> int

Number of steps in a path.

val of_path : Path.t -> t

Create a zigzag from a path.

val map : ( term -> term ) -> ( Step.t -> Step.t ) -> t -> t

Apply a context function to a path. In need to have two function because of typing issues (variance and polymorphic variants...), but they will always be the same in practice.

val subst : Subst.t -> t -> t

Apply a substitution.

val rule_occurences : Rule.t -> t -> int

Number of occurences of a given rule in a path.

val rule_algebraic_occurences : Rule.t -> t -> int

Number of occurences of a given rule in a path, counted negatively when inverted.

val is_id : t -> bool

val is_inv : t -> bool

val has_rule : Rule.t -> t -> bool

Whether a path contains a rule.

val canonize : t -> t

Put path in canonical form.

val value : Rule.t -> t -> t

Express a rule as a zigzag in a cell.

val replace_rule : Rule.t -> t -> t -> t

Replace a rule by a path in a path.

val nth_step : int -> t -> bool * Step.t

nth step in a path supposed to be in canonical form.

val nth_term : int -> t -> term

nth term in a path supposed to be in canonical form.

val to_list : t -> t list

List.of_steps in a path.

val parse : t -> string -> t