Module Term.RS

Rewriting systems.

val list_remove_nth : int -> 'a list -> 'a list * 'a list

module Rule : sig ... end

Rewriting rules.

type rule = Rule.t

type t = {

operations : Op.t list;

rules : rule list;

}

A rewriting system.

type rs = t

val operations : t -> Op.t list

val rules : t -> rule list

val filter : (rule -> bool) -> t -> t

Select rules satisfying a predicate.

val make : Op.t list -> rule list -> t

val empty : Op.t list -> t

Empty rewriting system.

val to_string : ?var:(unit -> Var.t -> string) -> t -> string

val find : t -> string -> Rule.t

Find the rule with given name.

module Step : sig ... end

Rewriting steps.

type step = Step.t

val steps : t -> t -> step list

All possible rewriting steps.

module Path : sig ... end

Rewriting paths.

val normalize : t -> term -> Path.t

Normalize a term.

val critical_rules : Rule.t -> Rule.t -> (Step.t * Step.t) list

Unify the source of r2 with a subterm of the source of r1.

val critical : t -> (Step.t * Step.t) list

Critical branchings.

val orient : gt:(term -> term -> bool) -> t -> t

Orient rules according to a particular order.

val knuth_bendix : ?gt:(term -> term -> bool) -> ?namer:(unit -> string) -> ?callback:(t -> unit) -> t -> t

Knuth-Bendix completion. gt is the strict order on terms, callback is a function which is called regularly with the current rewriting system as argument (useful to display during the completion).

exception Not_confluent

Raised when the system is not confluent.

val squier : t -> (Path.t * Path.t) list

Squier completion.

module Zigzag : sig ... end

Rewriting zigzags.

module Loop : sig ... end

Loops.

module Coherence : sig ... end

Coherence cells.

module Coherent : sig ... end

Coherent presentations.